% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ced-package.R
\docType{package}
\name{ced}
\alias{ced}
\alias{ced-package}
\title{The Compact Encoding Detector}
\description{
R bindings of the Google Compact Encoding Detection library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://artemklevtsov.gitlab.io/ced}
  \item \url{https://gitlab.com/artemklevtsov/ced}
  \item Report bugs at \url{https://gitlab.com/artemklevtsov/ced/issues}
}

}
\author{
\strong{Maintainer}: Artem Klevtsov \email{a.a.klevtsov@gmail.com} (\href{https://orcid.org/0000-0003-0492-6647}{ORCID})

Other contributors:
\itemize{
  \item Philipp Upravitelev \email{upravitelev@gmail.com} [contributor]
  \item Google Inc. [copyright holder]
}

}
